Updated Image.paste() with Image.alpha_composite
This was done so that the user's circular pfp wouldn't loose its transparency while being pasted on the background. Also Image.ANTIALIAS was done because of which the edges are much smoother now. 

The images used in Image.alpha_composite has to be RGBA so both images were converted to RGBA.

Exact position of the images need to be changed and fixed so as to fit the background.