added getserver()
server = client.get_guild() had to be used in every function to use the server's attributes. Now, it has been made global. Any function that needs the server id can get it using server=getserver() so that in order to change the server only the server id in getserver() function needs to be changed.